added test and put nystrom attention in the attention file
